backend/requirements.txt
# Django Core
Django>=5.0,<6.0

# API Framework
djangorestframework>=3.14

# PostgreSQL
psycopg[binary]>=3.1

# CORS for Next.js
django-cors-headers>=4.0

# Environment variables
python-dotenv>=1.0

# Background tasks (Celery)
celery>=5.3
redis>=5.0
django-celery-beat>=2.5

# Logging
structlog>=24.0

# Development
django-extensions>=3.2
ipython>=8.0
